+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Bugs Fixed
 * [Fix tooltips not working on grouped use statements](https://gitlab.com/php-integrator/core/issues/136)
+* [Using a trait in an interface will no longer crash the server](https://gitlab.com/php-integrator/core/issues/133)
 * [Fix unsupported meta file static method types throwing an error instead of being skipped](https://gitlab.com/php-integrator/core/issues/130)
 * [Project paths containing the tilde representing the home folder will now be properly expanded](https://gitlab.com/php-integrator/core/merge_requests/72)
 * [Fix folder scanning occurring twice during indexing, once for counting the total amount (for progress streaming) and once for actual indexing](https://github.com/php-integrator/atom-base/issues/314#issuecomment-320315228)